+++ components/gfx/platform/macos/font.rs
@@ -47,8 +47,8 @@ impl FontTable {
 }
 
 impl FontTableMethods for FontTable {
-    fn with_buffer<F>(&self, blk: F) where F: FnOnce(*const u8, uint) {
-        blk(self.data.bytes().as_ptr(), self.data.len() as uint);
+    fn with_buffer<F>(&self, blk: F) where F: FnOnce(*const u8, usize) {
+        blk(self.data.bytes().as_ptr(), self.data.len() as usize);
     }
 }
